+++ test/deprecate/test_deprecate_umphreys_mcgee.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require_relative '../../test_helper'
+require_relative '../test_helper'
 
-class TestFakerUmphreysMcgee < Test::Unit::TestCase
+class TestDeprecateUmphreysMcgee < Test::Unit::TestCase
   def setup
     @tester = Faker::UmphreysMcgee
   end